# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mystics Inventory API — multi-tenant inventory SaaS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /warehouses:
    get:
      operationId: listWarehouses
      tags: [warehouses]
      parameters:
        - in: query
          name: includeVirtual
          required: false
          schema: { type: boolean }
          description: When true, include virtual job-worker warehouses in the result. Defaults to false.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Warehouse" }
    post:
      operationId: createWarehouse
      tags: [warehouses]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWarehousePayload" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Warehouse" }
  /warehouses/{id}:
    get:
      operationId: getWarehouse
      tags: [warehouses]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Warehouse" }
    patch:
      operationId: updateWarehouse
      tags: [warehouses]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWarehousePayload" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Warehouse" }
    delete:
      operationId: deleteWarehouse
      tags: [warehouses]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "204": { description: No Content }
...
components:
  schemas:
...
    Warehouse:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        addressLine1: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        isDefault: { type: boolean }
        isVirtual: { type: boolean }
        jobWorkerSupplierId: { type: ["integer", "null"] }
        shopifyLocationId: { type: ["string", "null"] }
        shopifyLocationName: { type: ["string", "null"] }
        createdAt: { type: string }
      required:
        [
          id,
          name,
          code,
          isDefault,
          isVirtual,
          jobWorkerSupplierId,
          createdAt,
          addressLine1,
          city,
          state,
          country,
          shopifyLocationId,
          shopifyLocationName,
        ]
    CreateWarehousePayload:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        addressLine1: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        isDefault: { type: boolean }
      required: [name, code]
    UpdateWarehousePayload:
      type: object
      properties:
        name: { type: string }
        code: { type: string }
        addressLine1: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        isDefault: { type: boolean }
        shopifyLocationId: { type: ["string", "null"] }
...